Updated unit test to conform to new changes, removed state check from core constructor
